R: move classes to their own files
